clean up core summaries
